% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_teal4real.R
\name{use_teal4real}
\alias{use_teal4real}
\title{Install Teal4Real files}
\usage{
use_teal4real(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Should existing Teal4Real app files be overwritten?}
}
\description{
This function asks the user to choose a directory where to place the R files that make up the Teal4Real shiny app.
Take into consideration that in order to publish a Teal4Reall app to an rsconnect server, any R code required to generate
the analysis dataset should be either in the same folder as the Teal4Real files, linked into said folder using dynamic links,
or accessible via calling exported functions of a data generation package.
}
\details{
Of the copied files, those with "user" in the file name have to be configured by the user/programmer creating a Teal4Real app.
}
\examples{
\dontrun{
Teal4Real::use_teal4real()
}
}
